+++ tools/automation/__main__.py
@@ -10,7 +10,7 @@ import automation.clibuild
 import automation.clipublish
 import automation.style
 import automation.tests
-import automation.command_lint
+import automation.cli_linter
 
 
 def main():
@@ -22,7 +22,7 @@ def main():
     automation.clipublish.init_args(sub_parser)
     automation.style.init_args(sub_parser)
     automation.tests.init_args(sub_parser)
-    automation.command_lint.init_args(sub_parser)
+    automation.cli_linter.init_args(sub_parser)
 
     if sys.argv[1:]:
         args = parser.parse_args()